requirements.txt
tensorflow-gpu~=2.8.0
numpy~=1.22.4
matplotlib~=3.5.1
sklearn~=0.0
scikit-learn~=1.0.2
pandas~=1.4.1
xgboost~=1.6.1